css-align-3/default-alignment/place-items-shorthand-003.html

<!DOCTYPE html>
<title>CSS Box Alignment: place-items shorthand - initial value</title>
<link rel="author" title="Javier Fernandez" href="mailto:jfernandez@igalia.com" />
<link rel="help" href="http://www.w3.org/TR/css3-align/#place-items-property" />
<meta name="assert" content="Check that place-items's 'initial' value expands to 'align-items' and 'justify-items'." />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #test {
        align-items: start;
        justify-items: end;
    }
</style>
<div id="log"></div>
<div id="test"></div>
<script>
    var div = document.getElementById("test");
    div.setAttribute("style", "place-items: initial");

    test(function() {
        assert_equals(div.style["align-items"],
            "initial", "place-items specified value for align-items");
    }, "Check place-items: initial - align-items resolved value");

    test(function() {
        assert_equals(div.style["justify-items"],
            "initial", "place-items specified value for justify-items");
    }, "Check place-items: initial - justify-items resolved value");
</script>
